Last fixes sent by Robert.
His descricption:

You forgot to add SmartTag# to the SmartTagFunc() function for use
with 'tagfunc', to the test script, and in the docs.  I've done that and
made a few other changes.  I put some of the doc file back in the script
as it was more implementation details, and included a brief summary of
what the script's for in both places.  Also fixed a bug or two and added
relevant tests to the test script.

[.. including] the change to avoid the crash [ Lech was ] getting.